requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
pip~=21.2.3
click~=8.1.3
certifi~=2021.10.8
requests~=2.28.1
six~=1.16.0
python-dateutil~=2.8.2
idna~=3.3
urllib3~=1.26.12
setuptools~=57.4.0
telebot~=0.0.4
python-telegram-bot-calendar~=1.0.5
protobuf~=3.20.3
loguru~=0.6.0
httplib2~=0.22.0
google-api-python-client~=2.96.0
oauth2client~=4.1.3